+++ tide/wall/StreamImage.h
@@ -63,6 +63,9 @@ public:
     /** @copydoc Image::getData */
     const uint8_t* getData(uint texture) const final;
 
+    /** @copydoc Image::getRowOrder */
+    deflect::RowOrder getRowOrder() const final;
+
     /** @copydoc Image::getFormat */
     TextureFormat getFormat() const final;