+++ tide/wall/TextureNodeYUV.cpp
@@ -72,15 +72,19 @@ uniform lowp float qt_Opacity;
 uniform lowp sampler2D y_tex;
 uniform lowp sampler2D u_tex;
 uniform lowp sampler2D v_tex;
+uniform lowp bool reverse_orientation;
 varying vec2 vTexCoord;
 const vec3 offset = vec3(-0.0625, -0.5, -0.5);
 const vec3 R_cf = vec3(1.164383,  0.000000,  1.596027);
 const vec3 G_cf = vec3(1.164383, -0.391762, -0.812968);
 const vec3 B_cf = vec3(1.164383,  2.017232,  0.000000);
 void main() {
-  float y = texture2D(y_tex, vTexCoord).r;
-  float u = texture2D(u_tex, vTexCoord).r;
-  float v = texture2D(v_tex, vTexCoord).r;
+  vec2 texCoord = vTexCoord;
+  if(reverse_orientation)
+     texCoord.y = 1.0 - texCoord.y;
+  float y = texture2D(y_tex, texCoord).r;
+  float u = texture2D(u_tex, texCoord).r;
+  float v = texture2D(v_tex, texCoord).r;
   vec3 yuv = vec3(y, u, v);
   yuv += offset;
   float r = dot(yuv, R_cf);
@@ -100,6 +104,7 @@ struct YUVState
     std::unique_ptr<QSGTexture> textureU;
     std::unique_ptr<QSGTexture> textureV;
     TextureFormat textureFormat;
+    bool reverseOrientation = false;
 
     std::unique_ptr<QOpenGLBuffer> pboY;
     std::unique_ptr<QOpenGLBuffer> pboU;
@@ -137,6 +142,9 @@ public:
 
         gl->glActiveTexture(GL_TEXTURE0);
         newState->textureY->bind();
+
+        program()->setUniformValue("reverse_orientation",
+                                   (int)newState->reverseOrientation);
     }
 
     void resolveUniforms() final
@@ -210,6 +218,8 @@ void TextureNodeYUV::uploadTexture(const Image& image)
 
     _nextTextureSize = image.getTextureSize();
     _nextFormat = image.getFormat();
+    state->reverseOrientation =
+        image.getRowOrder() == deflect::RowOrder::bottom_up;
 }
 
 void TextureNodeYUV::swap()