+++ doc/Changelog.md
@@ -3,6 +3,8 @@ Changelog {#changelog}
 
 # Release 1.4 (git master)
 
+* [189](https://github.com/BlueBrain/Tide/issues/189):
+  Support for textures / PixelStreams in OpenGL (bottom-up) row order.
 * [179](https://github.com/BlueBrain/Tide/issues/179):
   Allow turning pixel stream compression on/off while streaming (bugfix).
 * [178](https://github.com/BlueBrain/Tide/issues/178):